[ 1954980 ] Wrong type for _CMPIBroker.brokerCapabilities

+++ cmpi/cmpift.h
@@ -225,7 +225,7 @@ extern "C" {
      /** 32 bits describing CMPI features supported by this CIMOM.
          See CMPI_MB_Class_x and CMPI_MB_Supports_xxx flags.
      */
-     unsigned long brokerClassification;
+     unsigned int brokerClassification;
      /** CIMOM version as defined by CIMOM
      */
      unsigned int brokerVersion;
